🧹chore: API폴더구조 개선
