+++ api/auth/index.ts
@@ -1,6 +1,19 @@
 import instance from "@/api/axios";
 import { ENDPOINTS } from "@/api/config";
 
+export const postLogin = async () => {
+  try {
+    const res = await instance.post(ENDPOINTS.AUTH.POST, {
+      email: "jieun@codeit.com",
+      password: "123asdf",
+    });
+
+    if (res.status === 200) return res.data;
+  } catch (error: any) {
+    console.error(error.response.data.message);
+  }
+};
+
 export const putPassword = async ({
   token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NjYyOTgsImlzcyI6InNwLXRhc2tpZnkifQ.zNaGd4uESNMzrDDHokuybQNJs_CkFLY7SpYKgafPBl0",
 }: {
@@ -19,9 +32,9 @@ export const putPassword = async ({
         },
       },
     );
-    console.log(res.data);
+
     if (res.status === 200) return res.data;
   } catch (error: any) {
-    console.log(error);
+    console.error(error.response.data.message);
   }
 };
