--- api/cards/getCard.ts
@@ -1,24 +1,0 @@
-import instance from "@/api/axios";
-import { ENDPOINTS } from "@/api/config";
-
-interface getCardProps {
-  cardId?: string;
-  token?: string;
-}
-
-export const getCard = async ({
-  cardId = "77",
-  token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NjYyOTgsImlzcyI6InNwLXRhc2tpZnkifQ.zNaGd4uESNMzrDDHokuybQNJs_CkFLY7SpYKgafPBl0",
-}: getCardProps) => {
-  try {
-    const res = await instance.get(ENDPOINTS.CARDS.GET_CARD(cardId), {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    console.log(res.data);
-    if (res.status === 200) return res.data.data;
-  } catch (error: any) {
-    console.log(error.response.data.message);
-  }
-};
