+++ api/comments/index.ts
@@ -1,0 +1,105 @@
+import instance from "@/api/axios";
+import { ENDPOINTS } from "@/api/config";
+
+interface deleteCommentsProps {
+  commentId?: string;
+  token?: string;
+}
+
+export const deleteComments = async ({
+  commentId = "98",
+  token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NjYyOTgsImlzcyI6InNwLXRhc2tpZnkifQ.zNaGd4uESNMzrDDHokuybQNJs_CkFLY7SpYKgafPBl0",
+}: deleteCommentsProps) => {
+  try {
+    const res = await instance.delete(ENDPOINTS.COMMENT.DELETE(commentId), {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+  } catch (error: any) {
+    console.error(error.response.data.message);
+  }
+};
+
+interface getCommentsProps {
+  cardId?: number;
+  size?: number;
+  cursorId?: number;
+  token?: string;
+}
+
+export const getComments = async ({
+  cardId = 4,
+  size = 5,
+  cursorId,
+  token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NjYyOTgsImlzcyI6InNwLXRhc2tpZnkifQ.zNaGd4uESNMzrDDHokuybQNJs_CkFLY7SpYKgafPBl0",
+}: getCommentsProps) => {
+  try {
+    const res = await instance.get(ENDPOINTS.COMMENT.GET, {
+      params: {
+        size,
+        cursorId,
+        cardId,
+      },
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    if (res.status === 200) return res.data.data;
+  } catch (error: any) {
+    console.error(error.response.data.message);
+  }
+};
+
+export const postComments = async ({
+  token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NjYyOTgsImlzcyI6InNwLXRhc2tpZnkifQ.zNaGd4uESNMzrDDHokuybQNJs_CkFLY7SpYKgafPBl0",
+}: {
+  token: string;
+}) => {
+  try {
+    const res = await instance.post(
+      ENDPOINTS.COMMENT.POST,
+      {
+        content: "좋은 하루",
+        cardId: 4,
+        columnId: 16,
+        dashboardId: 5,
+      },
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      },
+    );
+    if (res.status === 200) return res.data;
+  } catch (error: any) {
+    console.error(error.response.data.message);
+  }
+};
+
+interface putCommentsProps {
+  commentId: string;
+  token: string;
+}
+
+export const putComments = async ({
+  commentId = "98",
+  token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NjYyOTgsImlzcyI6InNwLXRhc2tpZnkifQ.zNaGd4uESNMzrDDHokuybQNJs_CkFLY7SpYKgafPBl0",
+}: putCommentsProps) => {
+  try {
+    const res = await instance.put(
+      ENDPOINTS.COMMENT.PUT(commentId),
+      {
+        content: "나은님 메롱",
+      },
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      },
+    );
+    if (res.status === 200) return res.data;
+  } catch (error: any) {
+    console.error(error.response.data.message);
+  }
+};
