+++ api/members/index.ts
@@ -25,9 +25,29 @@ export const getMembers = async ({
         Authorization: `Bearer ${token}`,
       },
     });
-    console.log(res.data);
     if (res.status === 200) return res.data.data;
   } catch (error: any) {
-    console.log(error.response.data.message);
+    console.error(error.response.data.message);
+  }
+};
+
+interface deleteMembersProps {
+  memberId: string;
+  token?: string;
+}
+
+export const deleteMembers = async ({
+  memberId = "241",
+  token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTMsInRlYW1JZCI6IjEtMDgiLCJpYXQiOjE3MDM1NzU1MjgsImlzcyI6InNwLXRhc2tpZnkifQ.vPTurAcm35kevcT9alVW2SxsjFcaKqnmd_mpgVwWfRU",
+}: deleteMembersProps) => {
+  try {
+    const res = await instance.delete(ENDPOINTS.MEMBERS.DELETE(memberId), {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    // if (res.status === 200) return res.data.data;
+  } catch (error: any) {
+    console.error(error.response.data.message);
   }
 };
